pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "APIProject"
version = "0.1.0"
authors = [{ name="M.Dikiy"}]
dependencies = [
    "allure-pytest==2.13.5",
    "allure-python-commons==2.13.5",
    "attrs==24.2.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "colorama==0.4.6",
    "deepdiff==8.0.1",
    "greenlet==3.1.1",
    "idna==3.10",
    "iniconfig==2.0.0",
    "jsonpath-ng==1.7.0",
    "orderly-set==5.2.2",
    "packaging==24.2",
    "pluggy==1.5.0",
    "ply==3.11",
    "psycopg2==2.9.10",
    "pytest==8.3.3",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "SQLAlchemy==2.0.36",
    "typing_extensions==4.12.2",
    "urllib3==2.2.3"
]

[tool.pytest.ini_options]
addopts = "--alluredir=allure-results"